renaming all reducer routes and adding page logic to createworkitem
